requirements.txt
waitress
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-bcrypt
email-validator
phonenumberslite
pillow
openpyxl
